Split Main Class into multiple Classes
